refactor(cli): Reduce unnecessary use of async + reorganize methods
