+++ cli/tracecat_cli/__main__.py
