+++ cli/tracecat_cli/_utils.py
@@ -1,28 +1,8 @@
 from pathlib import Path
 
-import httpx
 import orjson
-import rich
 import typer
 from rich.table import Table
-
-from ._config import config
-
-
-def user_client() -> httpx.AsyncClient:
-    """Returns an asynchronous httpx client with the user's JWT token."""
-    return httpx.AsyncClient(
-        headers={"Authorization": f"Bearer {config.jwt_token}"},
-        base_url=config.api_url,
-    )
-
-
-def user_client_sync() -> httpx.Client:
-    """Returns a synchronous httpx client with the user's JWT token."""
-    return httpx.Client(
-        headers={"Authorization": f"Bearer {config.jwt_token}"},
-        base_url=config.api_url,
-    )
 
 
 def dynamic_table(data: list[dict[str, str]], title: str) -> Table:
@@ -56,11 +36,3 @@ def read_input(data: str) -> dict[str, str]:
         return orjson.loads(data)
     except orjson.JSONDecodeError as e:
         raise typer.BadParameter(f"Invalid JSON: {e}") from e
-
-
-def handle_response(res: httpx.Response) -> None:
-    if res.status_code == 422:
-        rich.print("[red]Data validation error[/red]")
-        rich.print(res.json())
-        raise typer.Exit()
-    res.raise_for_status()
